refactor: retira a logica de gráficos para um arquivo a parte, para deixar o componente mais legivel
